html/semantics/tabular-data/the-table-element/border.html

<!DOCTYPE HTML>
<html>
<head>
	<title>The border attribute</title>
	<link rel="author" title="Erika Navara" href="mailto:edoyle@microsoft.com">
	<link rel="help" href="http://www.w3.org/TR/html5/tabular-data.html#the-table-element" />
	<link rel="help" href="http://www.w3.org/TR/html5/tabular-data.html#dom-table-border" />
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
</head>
<body>
	<div id="log"></div>
	<table id="table1" style="display:none">
		<tr>
			<td>cell</td>
			<td>cell</td>
		</tr>
	</table>
	<table id="table2" border="" style="display:none">
		<tr>
			<td>cell</td>
			<td>cell</td>
		</tr>
	</table>
	<table id="table3" border="1" style="display:none">
		<tr>
			<td>cell</td>
			<td>cell</td>
		</tr>
	</table>
	<script>
		test(function () {
			assert_idl_attribute(table1, "border");
			assert_idl_attribute(table2, "border");
			assert_idl_attribute(table3, "border");
			assert_equals(table1.border, "");
			assert_equals(table2.border, "");
			assert_equals(table3.border, "1");
		}, "The border IDL attribute reflects the content attribute of the same name.")
	</script>
</body>
</html>
